Server/pip_requirements.txt
gevent
django
gevent-websocket
gevent-socketio
